refactor(core): Move queue recovery to scaling service (no-changelog)
