+++ packages/cli/src/scaling/__tests__/scaling.service.test.ts
@@ -7,6 +7,9 @@ import type { Job, JobData, JobOptions, JobQueue } from '../types';
 import { ApplicationError } from 'n8n-workflow';
 import { mockInstance } from '@test/mocking';
 import { GlobalConfig } from '@n8n/config';
+import { InstanceSettings } from 'n8n-core';
+import type { OrchestrationService } from '@/services/orchestration.service';
+import Container from 'typedi';
 
 const queue = mock<JobQueue>({
 	client: { ping: jest.fn() },
@@ -33,9 +36,26 @@ describe('ScalingService', () => {
 		},
 	});
 
+	const instanceSettings = Container.get(InstanceSettings);
+	let orchestrationService = mock<OrchestrationService>({ isMultiMainSetupEnabled: false });
+	let scalingService: ScalingService;
+
 	beforeEach(() => {
 		jest.clearAllMocks();
 		config.set('generic.instanceType', 'main');
+		scalingService = new ScalingService(
+			mock(),
+			mock(),
+			mock(),
+			globalConfig,
+			mock(),
+			instanceSettings,
+			orchestrationService,
+		);
+	});
+
+	afterEach(() => {
+		scalingService.stopQueueRecovery();
 	});
 
 	describe('setupQueue', () => {
@@ -43,7 +63,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			const { prefix, settings } = globalConfig.queue.bull;
 			const Bull = jest.mocked(BullModule.default);
 
@@ -71,7 +90,15 @@ describe('ScalingService', () => {
 			 * Arrange
 			 */
 			config.set('generic.instanceType', 'worker');
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
+			const scalingService = new ScalingService(
+				mock(),
+				mock(),
+				mock(),
+				globalConfig,
+				mock(),
+				instanceSettings,
+				orchestrationService,
+			);
 			await scalingService.setupQueue();
 			const concurrency = 5;
 
@@ -90,7 +117,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
 
 			/**
@@ -100,13 +126,13 @@ describe('ScalingService', () => {
 		});
 	});
 
-	describe('pauseQueue', () => {
-		it('should pause the queue', async () => {
+	describe('stop', () => {
+		it('should pause the queue and stop queue recovery', async () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
+			const stopQueueRecoverySpy = jest.spyOn(scalingService, 'stopQueueRecovery');
 
 			/**
 			 * Act
@@ -117,6 +143,7 @@ describe('ScalingService', () => {
 			 * Assert
 			 */
 			expect(queue.pause).toHaveBeenCalledWith(true, true);
+			expect(stopQueueRecoverySpy).toHaveBeenCalled();
 		});
 	});
 
@@ -125,7 +152,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
 
 			/**
@@ -145,7 +171,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
 			queue.add.mockResolvedValue(mock<Job>({ id: '456' }));
 
@@ -168,7 +193,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
 			const jobId = '123';
 			queue.getJob.mockResolvedValue(mock<Job>({ id: jobId }));
@@ -191,7 +215,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
 			queue.getJobs.mockResolvedValue([mock<Job>({ id: '123' })]);
 
@@ -212,7 +235,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
 			// @ts-expect-error - Untyped but possible Redis response
 			queue.getJobs.mockResolvedValue([mock<Job>(), null]);
@@ -234,7 +256,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
 			const job = mock<Job>({ isActive: jest.fn().mockResolvedValue(true) });
 
@@ -254,7 +275,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
 			const job = mock<Job>({ isActive: jest.fn().mockResolvedValue(false) });
 
@@ -274,7 +294,6 @@ describe('ScalingService', () => {
 			/**
 			 * Arrange
 			 */
-			const scalingService = new ScalingService(mock(), mock(), mock(), globalConfig);
 			await scalingService.setupQueue();
 			const job = mock<Job>({
 				isActive: jest.fn().mockImplementation(() => {
@@ -291,6 +310,44 @@ describe('ScalingService', () => {
 			 * Assert
 			 */
 			expect(result).toBe(false);
+		});
+	});
+
+	describe('scheduleQueueRecovery', () => {
+		it('if leader, should schedule queue recovery', async () => {
+			/**
+			 * Arrange
+			 */
+			const scheduleSpy = jest.spyOn(scalingService, 'scheduleQueueRecovery');
+			instanceSettings.markAsLeader();
+
+			/**
+			 * Act
+			 */
+			await scalingService.setupQueue();
+
+			/**
+			 * Assert
+			 */
+			expect(scheduleSpy).toHaveBeenCalled();
+		});
+
+		it('if follower, should not schedule queue recovery', async () => {
+			/**
+			 * Arrange
+			 */
+			const scheduleSpy = jest.spyOn(scalingService, 'scheduleQueueRecovery');
+			instanceSettings.markAsFollower();
+
+			/**
+			 * Act
+			 */
+			await scalingService.setupQueue();
+
+			/**
+			 * Assert
+			 */
+			expect(scheduleSpy).not.toHaveBeenCalled();
 		});
 	});
 });
